pyproject.toml
[tool.poetry]
name = "sql-helper"
version = "0.1.0"
description = ""
authors = ["Martin Tomasi <martin.tomasi@gmail.com>"]
readme = "README.md"
packages = [{include = "sql_helper"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
langchain = "^0.0.258"
psycopg = {extras = ["binary"], version = "^3.1.10"}
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
